working navigation with design on home page
