--- SimpleApp/app.json
@@ -1,5 +1,0 @@
-{
-  "expo": {
-    "sdkVersion": "21.0.0"
-  }
-}
