+++ native-app/.gitignore
@@ -1,4 +1,3 @@
 node_modules/
 .expo/
 npm-debug.*
-package-lock.json
