+++ native-app/App.js
@@ -1,76 +1,54 @@
-import React, {Component} from 'react';
-import {StyleSheet, View, Text, AppRegistry, Button} from 'react-native';
-import Schedule from './containers/schedule/Schedule'
-import { StackNavigator } from "react-navigation";
+import React from 'react';
+import { StyleSheet, Text, View, TouchableOpacity } from 'react-native';
+import { StackNavigator } from 'react-navigation';
+import ScheduleScreen from './containers/ScheduleScreen'
+import TodoScreen from './containers/TodoScreen'
+import NotesScreen from './containers/NotesScreen'
 
-export default class App extends React.Component {
-  render() {
-    const {navigate} = this.props;
-    return (
-        <View style={styles.content}>
-          <View style={styles.headerContainer}>
-            <Text style={styles.text}>SaveMyLife.com</Text>
-          </View>
-          <View style={styles.back}>
-              <View style={styles.scheduleContainer}>
-                  <Button style={styles.scheduleBoxHeader} onPress={() => navigate('Schedule')} title={'test'}>
-                      <Text style={styles.innerText}>Schedule</Text>
-                  </Button>
-              </View>
-              <View style={styles.todoContainer}>
-                  <View style={styles.todoBoxHeader}>
-                      <Text style={styles.innerText}>2do</Text>
-                  </View>
-              </View>
-              <View style={styles.notesContainer}>
-                  <View style={styles.notesBoxHeader}>
-                      <Text style={styles.innerText}>Notes</Text>
-                  </View>
-              </View>
-          </View>
-        </View>
-    );
-  }
+class HomeScreen extends React.Component {
+    static navigationOptions = {
+        title: 'SaveMyLife.com',
+        headerStyle: { backgroundColor: '#212a34', elevation: 6,},
+        headerTitleStyle: { alignSelf: 'center', color: '#fff', paddingTop: 5},
+    };
+    render() {
+        const { navigate } = this.props.navigation;
+        return (
+            <View style={styles.container}>
+                <TouchableOpacity onPress={() => navigate('Schedule')} style={styles.schedule}>
+                    <Text style={styles.buttonText}>Schedule</Text>
+                </TouchableOpacity>
+                <TouchableOpacity onPress={() => navigate('Todo')} style={styles.todo}>
+                    <Text style={styles.buttonText}>Todo</Text>
+                </TouchableOpacity>
+                <TouchableOpacity onPress={() => navigate('Notes')} style={styles.notes}>
+                    <Text style={styles.buttonText}>Note</Text>
+                </TouchableOpacity>
+            </View>
+        );
+    }
 }
 
-const ModalStack = StackNavigator({
-    Schedule: {
-        screen: Schedule,
-    },
+const NativeApp = StackNavigator({
+    Home: { screen: HomeScreen },
+    Schedule: { screen: ScheduleScreen },
+    Todo: { screen: TodoScreen },
+    Notes: { screen: NotesScreen }
 });
 
+export default class App extends React.Component {
+    render() {
+        return <NativeApp />;
+    }
+}
 
 const styles = StyleSheet.create({
-    content: {
+    container: {
         flex: 1,
-    },
-
-    headerContainer: {
-        backgroundColor: '#212a34',
-        elevation: 6,
-        paddingTop: 30,
-        paddingBottom: 10,
-        shadowColor: 'rgba(0,0,0,0.3)',
-        shadowOpacity: 3,
-        shadowOffset: {
-            height: 0,
-            width: 1
-        },
-    },
-
-    text: {
-        color: '#fff',
-        textAlign: 'center',
-        fontFamily: 'sans-serif',
-        fontWeight: 'bold'
-    },
-
-    back: {
         backgroundColor: "#666",
-        flex: 1
     },
 
-    scheduleContainer: {
+    schedule: {
         backgroundColor: "#cfc",
         marginLeft: '3%',
         marginRight: '3%',
@@ -88,7 +66,7 @@ const styles = StyleSheet.create({
         alignItems: 'center',
     },
 
-    todoContainer: {
+    todo: {
         backgroundColor: "#ccf",
         marginLeft: '3%',
         marginRight: '3%',
@@ -106,7 +84,7 @@ const styles = StyleSheet.create({
         alignItems: 'center',
     },
 
-    notesContainer: {
+    notes: {
         backgroundColor: "#FFFFA5",
         marginLeft: '3%',
         marginRight: '3%',
@@ -125,10 +103,8 @@ const styles = StyleSheet.create({
         alignItems: 'center',
     },
 
-    innerText: {
+    buttonText: {
         fontSize: 40,
         color: '#212a34'
     }
 });
-
-AppRegistry.registerComponent('App', () => ModalStack);
