+++ native-app/containers/NotesScreen.js
