+++ native-app/containers/ScheduleScreen.js
