+++ native-app/package.json
@@ -20,7 +20,7 @@
   },
   "dependencies": {
     "expo": "^21.0.0",
-    "react": "^16.0.0-alpha.12",
+    "react": "16.0.0-alpha.12",
     "react-native": "^0.48.4",
     "react-navigation": "^1.0.0-beta.13"
   }
